- Updated reporting engine - Fixed a bug related to SMTP email configuration.

+++ src/Libraries/Frapid.Messaging/Smtp/Processor.cs
@@ -127,7 +127,7 @@ namespace Frapid.Messaging.Smtp
                             item?.Dispose();
                         }
 
-                        if (deleteAttachmentes)
+                        if (deleteAttachmentes && email.Status == Status.Completed)
                         {
                             this.DeleteFiles(attachments);
                         }
